DIV-6496: Using classes moved to common library
