+++ build.gradle
@@ -36,7 +36,7 @@ def versions = [
         commonsIo                    : '2.6',
         commonsLang3                 : '3.9',
         commonsMath3                 : '3.6.1',
-        divCommonLib                 : '1.1.27',
+        divCommonLib                 : '1.1.28',
         dumbster                     : '1.7.1',
         elasticsearch                : '7.9.0',
         feignHttpClient              : '10.3.0',
@@ -107,7 +107,6 @@ allprojects {
     }
 
     repositories {
-        mavenLocal()
         maven { url "https://dl.bintray.com/hmcts/hmcts-maven" }
         maven { url 'https://repo.spring.io/libs-milestone' }
         mavenCentral()
