+++ src/integrationTest/java/uk/gov/hmcts/reform/divorce/callback/ProcessPbaPaymentTest.java
@@ -4,8 +4,8 @@ import io.restassured.response.Response;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Value;
 import uk.gov.hmcts.reform.divorce.context.IntegrationTest;
+import uk.gov.hmcts.reform.divorce.model.ccd.Document;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CollectionMember;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.Document;
 
 import java.util.List;
 import java.util.Map;
