+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/domain/model/ccd/CoreCaseData.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import uk.gov.hmcts.reform.divorce.model.ccd.Document;
 import uk.gov.hmcts.reform.divorce.model.ccd.OrderSummary;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.payment.PaymentCollection;
