--- src/main/java/uk/gov/hmcts/reform/divorce/orchestration/domain/model/ccd/DivorceGeneralOrder.java
@@ -1,18 +1,0 @@
-package uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@Builder
-public class DivorceGeneralOrder {
-
-    @JsonProperty("Document")
-    private Document document;
-
-    @JsonProperty("GeneralOrderParties")
-    private List<GeneralOrderParty> generalOrderParties;
-}
