--- src/main/java/uk/gov/hmcts/reform/divorce/orchestration/domain/model/ccd/DocumentLink.java
@@ -1,25 +1,0 @@
-package uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class DocumentLink {
-
-    @JsonProperty("document_url")
-    private String documentUrl;
-
-    @JsonProperty("document_binary_url")
-    private String documentBinaryUrl;
-
-    @JsonProperty("document_filename")
-    private String documentFilename;
-}
