+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/AosService.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.divorce.orchestration.service;
 
+import uk.gov.hmcts.reform.divorce.model.parties.DivorceParty;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.parties.DivorceParty;
 
 import java.util.Map;
