+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/util/mapper/CcdMappers.java
@@ -1,8 +1,8 @@
 package uk.gov.hmcts.reform.divorce.orchestration.util.mapper;
 
+import uk.gov.hmcts.reform.divorce.model.ccd.Document;
+import uk.gov.hmcts.reform.divorce.model.ccd.DocumentLink;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CollectionMember;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.Document;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.DocumentLink;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.documentgeneration.GeneratedDocumentInfo;
 
 public class CcdMappers {
