+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/testutil/CaseDataTestHelper.java
@@ -1,9 +1,9 @@
 package uk.gov.hmcts.reform.divorce.orchestration.testutil;
 
 import com.fasterxml.jackson.core.type.TypeReference;
+import uk.gov.hmcts.reform.divorce.model.ccd.Document;
+import uk.gov.hmcts.reform.divorce.model.ccd.DocumentLink;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CollectionMember;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.Document;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.DocumentLink;
 
 import java.util.Map;
